refactor(types): replace generic primitives.ts with purpose-named files
primitives.ts was a generic bucket holding three unrelated concerns. Split
into speaking names: metrics.ts (Metrics), source-code.ts (FilePath,
FileContent, CodeLocation, CodeSnippet), session.ts (SessionId). Barrel
re-exports them, so the package's public type API is unchanged (these types
are exported via src/index.ts for external consumers — they have no internal
callers, which is expected for public-API vocabulary).

No behavior change: build clean, 2619/2619 tests pass, lint clean.
