+++ src/shared/types/session.ts
@@ -1,0 +1,2 @@
+/** Identifier for a single pipeline run/session. */
+export type SessionId = string;
