+++ src/shared/types/source-code.ts
@@ -1,7 +1,6 @@
-/** Generic, domain-agnostic value shapes shared across the pipeline. */
+/** File paths, file content, and positions/snippets within source code. */
 
 export type FilePath = string;
-export type SessionId = string;
 
 export type FileContent = {
   path: FilePath;
@@ -21,14 +20,4 @@ export type CodeLocation = {
 export type CodeSnippet = CodeLocation & {
   code: string;
   language?: string;
-};
-
-export type Metrics = {
-  duration: number;
-  filesAnalyzed: number;
-  issuesFound: number;
-  fixesApplied: number;
-  tokensUsed?: number;
-  cacheHits?: number;
-  cacheMisses?: number;
 };
